Fixing remaining circular dependencies. These are shown in Ivy partial compilation mode
