+++ projects/jsf/src/lib/form-content/tabs/tab/tab.component.ts
@@ -1,13 +1,13 @@
-import { Component, ContentChild, Input } from '@angular/core';
+import { Component, Input } from '@angular/core';
+import { FormGroup } from '@angular/forms';
 import { ParentDataItem } from '../../../models/parent-data-item';
-import { FormContentComponent } from '../../form-content.component';
 
 @Component({
   selector: 'jsf-tab',
   templateUrl: './tab.component.html'
 })
 export class TabComponent {
-  @ContentChild(FormContentComponent, { static: true}) public content: FormContentComponent;
   @Input() dataItem: ParentDataItem;
   @Input() selected = false;
+  @Input() formGroup: FormGroup;
 }
