+++ projects/jsf/src/lib/form-content/tabs/tabs.component.html
@@ -1,17 +1,22 @@
-<jsf-tabstrip>
-  <jsf-tab
-    *ngFor="let item of items; let i = index"
-    [selected]="i === 0"
-    [dataItem]="item">
-    <jsf-form-content
-      [formGroup]="getFormGroup(item)"
-      [formItems]="item.items"
-      [config]="config"
-      [templates]="templates"
-      [labelLengthClass]="getLabelLengthClass(item)"
-      (buttonEvent)="onButtonEvent($event)"
-      (templateEvent)="onTemplateEvent($event)"
-      [isEdit]="isEdit">
-    </jsf-form-content>
-  </jsf-tab>
-</jsf-tabstrip>
+<div class="tab-strip panel panel-default panel-tabbed">
+  <div class="panel-heading">
+    <div class="panel-title">
+      {{title}}
+    </div>
+    <ul class="nav nav-tabs">
+      <li class="clickable" *ngFor="let tab of tabs; let i = index" (click)="onTabClicked(tab)" [class.active]="tab.selected" [class.has-error]="!isValid(tab) && !isDisabled(tab)">
+        <a>{{tab.dataItem.label}}<span class="required-star" *ngIf="tabHasRequiredFields(tab)">*</span></a>
+      </li>
+    </ul>
+    <div *ngIf="selectedTab?.dataItem?.description" class="tab-description">
+      <span>{{selectedTab.dataItem.description}}</span>
+      <img *ngIf="selectedTab.dataItem.helpText"
+           class="help"
+           [title]="selectedTab.dataItem.helpText"
+           [src]="'assets/jsf-images/info.svg'">
+    </div>
+  </div>
+  <div class="panel-body">
+    <ng-content></ng-content>
+  </div>
+</div>
