--- projects/jsf/src/lib/form-content/tabs/tabstrip/tabstrip.component.scss
@@ -1,22 +1,0 @@
-@import '../../form-controls/form-control.component';
-
-option {
-  color: #000000
-}
-
-.default-option, .default-selected {
-  color: #808080;
-}
-
-.tab-description {
-  margin: 20px 10px 10px 0px
-}
-
-.has-error {
-  border-bottom: 1px solid #ff0000;
-}
-
-.required-star {
-  color: #f00;
-  margin-left: 5px;
-}
